refactor : package구조 개선 person(ddl) -> sql/entity 디렉토리로 이동
