+++ src/main/java/persistence/sql/DatabaseDialect.java
@@ -1,4 +1,4 @@
-package persistence.sql.ddl;
+package persistence.sql;
 
 public enum DatabaseDialect {
     MYSQL,
